+++ lib/gruf/client.rb
@@ -15,9 +15,6 @@
 # COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
 # OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
-require_relative 'client/error'
-require_relative 'client/error_factory'
-
 module Gruf
   ##
   # Abstracts out the calling interface for interacting with gRPC clients. Streamlines calling and provides
