+++ lib/gruf/client/errors.rb
@@ -1,5 +1,5 @@
-#!/usr/bin/env ruby
-# coding: utf-8
+# frozen_string_literal: true
+
 # Copyright (c) 2017-present, BigCommerce Pty. Ltd. All rights reserved
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
@@ -15,11 +15,34 @@
 # COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
 # OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
-require 'bundler/setup'
-require 'gruf'
+module Gruf
+  class Client < SimpleDelegator
+    ##
+    # See https://github.com/grpc/grpc-go/blob/master/codes/codes.go for a detailed summary of each error type
+    #
+    module Errors
+      class Base < Gruf::Client::Error; end
+      class Error < Base; end
+      class Validation < Base; end
 
-$LOAD_PATH.unshift File.expand_path('../../spec/pb', __FILE__)
-require File.realpath("#{File.dirname(File.dirname(__FILE__))}/spec/support/grpc.rb")
+      class Ok < Base; end
 
-require 'irb'
-IRB.start
+      class InvalidArgument < Validation; end
+      class NotFound < Validation; end
+      class AlreadyExists < Validation; end
+      class OutOfRange < Validation; end
+
+      class Cancelled < Error; end
+      class DataLoss < Error; end
+      class DeadlineExceeded < Error; end
+      class FailedPrecondition < Error; end
+      class Internal < Error; end
+      class PermissionDenied < Error; end
+      class ResourceExhausted < Error; end
+      class Unauthenticated < Error; end
+      class Unavailable < Error; end
+      class Unimplemented < Error; end
+      class Unknown < Error; end
+    end
+  end
+end
