+++ lib/gruf/error.rb
@@ -15,10 +15,6 @@
 # COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
 # OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
-require_relative 'errors/field'
-require_relative 'errors/debug_info'
-require_relative 'serializers/errors/base'
-require_relative 'serializers/errors/json'
 
 module Gruf
   ##
