+++ lib/gruf/interceptors/instrumentation/request_logging/interceptor.rb
@@ -16,9 +16,6 @@
 # OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #
 require 'socket'
-require_relative 'formatters/base'
-require_relative 'formatters/logstash'
-require_relative 'formatters/plain'
 
 module Gruf
   module Interceptors
